html/browsers/browsing-the-web/read-plugin/load-plugin.html

<!DOCTYPE html>
<title>Page load processing model for plugins</title>
<link rel="author" href="masonf@chromium.org">
<link rel="help" href="https://html.spec.whatwg.org/#read-plugin">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>

<iframe src="support/sandbox.pdf"></iframe>

<script>
const t = async_test("Checking document metadata for PDF file");
const iframe = document.querySelector('iframe');
iframe.onload = function() {
  var doc = iframe.contentDocument;
  t.step(function() {
    if (doc === null) {
      // Some UA's render PDF as "inline content that doesn't have a DOM", which is ok
      t.done();
    } else {
      // Otherwise, it should be "content that uses plugins" and the document should
      // be in standards mode.
      assert_equals(doc.compatMode, "CSS1Compat");
      assert_equals(doc.contentType, "application/pdf");
      t.done();
    }
  })
};
</script>
